# Conflicts:
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/actualworkinghours/TotalWorkingTime.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/daily/vacationusetime/AnnualOfDaily.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/CalculateDailyRecordServiceImpl.java
#	nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/dailyprocess/calc/OverTimeSheet.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/workrecord/monthcal/company/DelComMonthCalSetCommandHandler.java
@@ -4,13 +4,18 @@
  *****************************************************************/
 package nts.uk.ctx.at.record.app.command.workrecord.monthcal.company;
 
+import java.util.Optional;
+
 import javax.ejb.Stateless;
 import javax.inject.Inject;
 
 import nts.arc.layer.app.command.CommandHandler;
 import nts.arc.layer.app.command.CommandHandlerContext;
+import nts.uk.ctx.at.record.dom.workrecord.monthcal.company.ComDeforLaborMonthActCalSet;
 import nts.uk.ctx.at.record.dom.workrecord.monthcal.company.ComDeforLaborMonthActCalSetRepository;
+import nts.uk.ctx.at.record.dom.workrecord.monthcal.company.ComFlexMonthActCalSet;
 import nts.uk.ctx.at.record.dom.workrecord.monthcal.company.ComFlexMonthActCalSetRepository;
+import nts.uk.ctx.at.record.dom.workrecord.monthcal.company.ComRegulaMonthActCalSet;
 import nts.uk.ctx.at.record.dom.workrecord.monthcal.company.ComRegulaMonthActCalSetRepository;
 import nts.uk.shr.com.context.AppContexts;
 
@@ -42,9 +47,26 @@ public class DelComMonthCalSetCommandHandler extends CommandHandler<DelComMonthC
 	@Override
 	protected void handle(CommandHandlerContext<DelComMonthCalSetCommand> context) {
 		String cid = AppContexts.user().companyId();
-		comDeforLaborMonthActCalSetRepo.remove(cid);
-		comFlexMonthActCalSetRepo.remove(cid);
-		comRegulaMonthActCalSetRepo.remove(cid);
+		
+		Optional<ComDeforLaborMonthActCalSet> optComDeforLaborMonthActCalSet = comDeforLaborMonthActCalSetRepo
+				.find(cid);
+		if (optComDeforLaborMonthActCalSet.isPresent()) {
+			comDeforLaborMonthActCalSetRepo.remove(cid);
+		}
+		
+		Optional<ComFlexMonthActCalSet> optComFlexMonthActCalSet = comFlexMonthActCalSetRepo.find(cid);
+
+		if (optComFlexMonthActCalSet.isPresent()) {
+			comFlexMonthActCalSetRepo.remove(cid);
+		}
+
+		Optional<ComRegulaMonthActCalSet> optComRegulaMonthActCalSet = comRegulaMonthActCalSetRepo
+				.find(cid);
+
+		if (optComRegulaMonthActCalSet.isPresent()) {
+			comRegulaMonthActCalSetRepo.remove(cid);
+		}
+	
 	}
 
 }
